pyproject.toml
[tool.poetry]
name = "s-raid"
version = "0.1.0"
description = ""
authors = ["K-Maxim <m.karpenko96@mail.ru>"]
readme = "README.md"
packages = [{include = "s_raid"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
